+++ libdispatch/ddispatch.c
@@ -12,6 +12,7 @@ See LICENSE.txt for license information.
 #include "ncoffsets.h"
 #include "ncpathmgr.h"
 #include "ncxml.h"
+#include "ncglobal.h"
 #include "nc4internal.h"
 
 /* Required for getcwd, other functions. */
@@ -73,6 +74,7 @@ NCDISPATCH_initialize(void)
 	    fprintf(stderr,"Cannot find a temp dir; using ./\n");
 	    tempdir = ".";
 	}
+	assert(tempdir != NULL);
 	globalstate->tempdir= strdup(tempdir);
     }
 
@@ -138,92 +140,14 @@ NCDISPATCH_finalize(void)
 #if defined(NETCDF_ENABLE_DAP4)
    ncxml_finalize();
 #endif
+
+#ifdef NETCDF_ENABLE_S3
+    status = NC_s3sdkfinalize();
+#endif
+
     NC_freeglobalstate(); /* should be one of the last things done */
     return status;
 }
-
-/**************************************************/
-/* Global State constants and state */
-
-/* The singleton global state object */
-static NCglobalstate* nc_globalstate = NULL;
-
-/* Forward */
-static int NC_createglobalstate(void);
-
-/** \defgroup global_state Global state functions. */
-/** \{
-
-\ingroup global_state
-*/
-
-/* NCglobal state management */
-
-static int
-NC_createglobalstate(void)
-{
-    int stat = NC_NOERR;
-    const char* tmp = NULL;
-    
-    if(nc_globalstate == NULL) {
-        nc_globalstate = calloc(1,sizeof(NCglobalstate));
-    }
-    /* Initialize struct pointers */
-    if((nc_globalstate->rcinfo = calloc(1,sizeof(struct NCRCinfo)))==NULL)
-            {stat = NC_ENOMEM; goto done;}
-    if((nc_globalstate->rcinfo->entries = nclistnew())==NULL)
-            {stat = NC_ENOMEM; goto done;}
-    if((nc_globalstate->rcinfo->s3profiles = nclistnew())==NULL)
-            {stat = NC_ENOMEM; goto done;}
-
-    /* Get environment variables */
-    if(getenv(NCRCENVIGNORE) != NULL)
-        nc_globalstate->rcinfo->ignore = 1;
-    tmp = getenv(NCRCENVRC);
-    if(tmp != NULL && strlen(tmp) > 0)
-        nc_globalstate->rcinfo->rcfile = strdup(tmp);
-    /* Initialize chunk cache defaults */
-    nc_globalstate->chunkcache.size = DEFAULT_CHUNK_CACHE_SIZE;		    /**< Default chunk cache size. */
-    nc_globalstate->chunkcache.nelems = DEFAULT_CHUNKS_IN_CACHE;	    /**< Default chunk cache number of elements. */
-    nc_globalstate->chunkcache.preemption = DEFAULT_CHUNK_CACHE_PREEMPTION; /**< Default chunk cache preemption. */
-    
-done:
-    return stat;
-}
-
-/* Get global state */
-NCglobalstate*
-NC_getglobalstate(void)
-{
-    if(nc_globalstate == NULL)
-        NC_createglobalstate();
-    return nc_globalstate;
-}
-
-void
-NC_freeglobalstate(void)
-{
-    if(nc_globalstate != NULL) {
-        nullfree(nc_globalstate->tempdir);
-        nullfree(nc_globalstate->home);
-        nullfree(nc_globalstate->cwd);
-	nullfree(nc_globalstate->aws.default_region);
-	nullfree(nc_globalstate->aws.config_file);
-	nullfree(nc_globalstate->aws.profile);
-	nullfree(nc_globalstate->aws.session_token);
-	nullfree(nc_globalstate->aws.access_key_id);
-	nullfree(nc_globalstate->aws.secret_access_key);
-        if(nc_globalstate->rcinfo) {
-	    NC_rcclear(nc_globalstate->rcinfo);
-	    free(nc_globalstate->rcinfo);
-	}
-	nclistfree(nc_globalstate->pluginpaths);
-	free(nc_globalstate);
-	nc_globalstate = NULL;
-    }
-}
-
-/** \} */
 
 /**************************************************/
 /** \defgroup atomic_types Atomic Type functions */