+++ libnczarr/zopen.c
@@ -188,4 +188,3 @@ done:
     ncurifree(uri);
     return ZUNTRACE(stat);
 }
-